updated from giac_1.2.1-7

+++ src/quater.cc
@@ -87,7 +87,7 @@ namespace giac {
   // If it is only irreducible, returns 2 and sets vmin to a primitive poly mod p if primitive is true
   int is_irreducible_primitive(const modpoly & v,const gen & p,modpoly & vmin,bool primitive,GIAC_CONTEXT){
     vmin=v;
-    int m=v.size()-1;
+    int m=int(v.size())-1;
     if (m<2)
       return 0; // setsizeerr(gettext("irreducibility: degree too short")+gen(v).print());
     gen gpm=pow(p,m);
@@ -116,7 +116,7 @@ namespace giac {
     // Primi: must not divide x^[(p^m-1)/d]-1 for any prime divisor d of p^m-1
     gen pm_minus_1=gpm-1;
     vecteur vp(pfacprem(pm_minus_1,true,context0));
-    int ntest=vp.size();
+    int ntest=int(vp.size());
     for (int i=0;i<ntest;i+=2){
       // Compute x^[(p^m-1)/d] mod v, mod p, is it 1?
       gen pm_d=(gpm-1)/vp[i];
@@ -248,7 +248,7 @@ namespace giac {
 	return galois_field(args,true,contextptr);
       v=*args._VECTptr;
     }
-    int s=v.size();
+    int s=int(v.size());
     if (s<1 || !is_integer(v[0]))
       return gensizeerr(contextptr);
     if (_isprime(v[0],contextptr)==0){
@@ -681,7 +681,7 @@ namespace giac {
   }
 
   gen galois_field::sqrt(GIAC_CONTEXT) const {
-    unsigned m=P._VECTptr->size()-1;
+    unsigned m=unsigned(P._VECTptr->size())-1;
     // K* has cardinal p^m-1, a has a sqrt iff a^((p^m-1)/2)==1
     environment env;
     env.modulo=p;
@@ -751,7 +751,7 @@ namespace giac {
   gen galois_field::makegen(int i) const {
     if (P.type!=_VECT || p.type!=_INT_)
       return gendimerr();
-    unsigned n=P._VECTptr->size()-1;
+    unsigned n=unsigned(P._VECTptr->size())-1;
     //    i += pow(p,int(n)).val;
     vecteur res;
     for (unsigned j=0;j<n;++j){
@@ -807,7 +807,7 @@ namespace giac {
 
   gen galois_field::rand (GIAC_CONTEXT) const {
     int c=p.val;
-    int m=P._VECTptr->size()-1;
+    int m=int(P._VECTptr->size())-1;
     vecteur v(m);
     for (int i=0;i<m;++i){
       if (c==2)
